third_party/blink/web_tests/external/wpt/webrtc/RTCRtpReceiver-getStats.https-expected.txt

This is a testharness.js-based test.
[FAIL] receiver.getStats() via addTransceiver should return stats report containing inbound-rtp stats
  assert_true: expected true got false
[FAIL] receiver.getStats() via addTrack should return stats report containing inbound-rtp stats
  assert_true: expected true got false
[FAIL] receiver.getStats() should work on a stopped transceiver but not have inbound-rtp objects
  assert_true: expected true got false
[FAIL] receiver.getStats() should work with a closed PeerConnection but not have inbound-rtp objects
  assert_true: expected true got false
[FAIL] receiver.getStats() should return stats report containing ICE candidate stats
  assert_true: expected true got false
Harness: the test ran to completion.
